Organized Migrations. Phrase 1. Closes #1058
Migrations has been hotfixed in order to prevent no upgrading with a column already existed or some changes have been made no by the migrations. more work to be done here.

+++ api/migrations/upgrades/20150705024605_DropDirectusIpWhitelist.php
@@ -5,7 +5,9 @@ class DropDirectusIpWhitelist extends Ruckusing_Migration_Base
 {
     public function up()
     {
-        $this->drop_table('directus_ip_whitelist');
+        if ($this->has_table('directus_ip_whitelist')) {
+            $this->drop_table('directus_ip_whitelist');
+        }
     }//up()
 
     public function down()
